refactor(dao): EventDaoImpl to EventDao
- delete static factory method ofProd and table field in EventDao
- remove EventDao interface
- create test stubs as children of EventDao
